添加了，管理员的登录注销逻辑，主页广告的管理工作

+++ SSM03-Tmall/src/main/java/cn/bps/controller/AdminProductController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
@@ -27,7 +28,7 @@ public class AdminProductController {
     @Autowired
     private ProductBindFilterService productBindFilterService;
     @Autowired
-    private ProductService  productService;
+    private ProductService productService;
     @Autowired
     private CategoryService categoryService;
     @Autowired
@@ -39,101 +40,118 @@ public class AdminProductController {
 
 
 
-
     @RequestMapping("")
     public String showProductView(Model model,
-                              @RequestParam(value = "start",defaultValue = "0")int start,
-                              @RequestParam(value = "key",defaultValue = "")String key){
+                                  @RequestParam(value = "start", defaultValue = "0") int start,
+                                  @RequestParam(value = "key", defaultValue = "") String key,
+                                  @ModelAttribute("sessionUsername") String username) {
+
+        if (username.equals("")) return "redirect:/admin/login";
 
         /*分页*/
-        Page page = new Page(start,8);
+        Page page = new Page(start, 8);
         Set<Integer> productIdSet = productBindFilterService.getAllProductIdSet();
         page.setTotal(productIdSet.size());
 
         Set<Integer> keyProductIdSet = productService.getProductIDSetByProductName(key);
 
-        if(!key.equals("")){ //如果设置了关键字检索
+        if (!key.equals("")) { //如果设置了关键字检索
             productIdSet.retainAll(keyProductIdSet);
         }
 
         /*开启下架产品*/
-        List<Product> products = productService.rowBoundsProduct(productIdSet, page.getStart(), page.getStep(),false);
-        model.addAttribute("products",products);
+        List<Product> products = productService.rowBoundsProduct(productIdSet, page.getStart(), page.getStep(), true); //下架商品可见
+        model.addAttribute("products", products);
 
-        Map<Integer,String> categoryMap =categoryService.getCategoryMap();
-        model.addAttribute("categoryMap",categoryMap);
+        Map<Integer, String> categoryMap = categoryService.getCategoryMap();
+        model.addAttribute("categoryMap", categoryMap);
 
         Map<Integer, String> images = productImageService.getImageUrls(products);
-        model.addAttribute("images",images);
+        model.addAttribute("images", images);
 
 
-
-        model.addAttribute("page",page);
+        model.addAttribute("page", page);
 
         return "admin/product/product";
     }
 
     @RequestMapping(value = "/add")
-    public String showProductAddingView(Model model){
+    public String showProductAddingView(Model model, @ModelAttribute("sessionUsername") String username) {
+
+        if (username.equals("")) return "redirect:/admin/login";
+
         List<Category> categories = categoryService.getCategories();
-        model.addAttribute("categories",categories);
+        model.addAttribute("categories", categories);
 
 
         List<FilterCase> filterCases = filterCaseService.getFilterList();
         List<Integer> filterIdList = filterCaseService.getFilterIdList();
         Map<Integer, List<ConcreteFilter>> filterMap = concreteFilterService.getFilterMap(filterIdList);
-        model.addAttribute("filterMap",filterMap);
-        model.addAttribute("filterCases",filterCases);
+        model.addAttribute("filterMap", filterMap);
+        model.addAttribute("filterCases", filterCases);
 
         return "admin/product/productAdd";
     }
+
     //提交修改的产品
     @RequestMapping(value = "/edit/post/{id}", method = RequestMethod.POST)
     public String postProduct(HttpServletRequest request,
-                              @PathVariable(value = "id")int id,
-                              @ModelAttribute cn.bps.domain.Product product)
-    {
+                              @PathVariable(value = "id") int id,
+                              @ModelAttribute cn.bps.domain.Product product,
+                              @ModelAttribute("sessionUsername") String username) {
 
+
+        if (username.equals("")) return "redirect:/admin/login";
+
+        //s商品信息上传
         MultipartFile multipartFile = product.getMultipartFile();
         String text = multipartFile.getOriginalFilename();
-
         String suffix = Util.matchSuffix(text);
-
         String fileName = Util.generatorRandomCode() + suffix;
-
         String filePath = null;
 
-        if(text.equals("")){
-            filePath = "http://temp.im/200x200";
-        }else{
-            filePath = "/img/file/"+fileName;
-            try{
-                File file = new File(request.getServletContext().getRealPath("/WEB-INF/img/file/"),fileName);
-                if(!file.exists()){
+        if (text.equals("")) {
+            filePath = "";
+        } else {
+            filePath = "/img/file/" + fileName;
+            try {
+                File file = new File(request.getServletContext().getRealPath("/WEB-INF/img/file/"), fileName);
+                if (!file.exists()) {
                     file.mkdir();
                 }
                 multipartFile.transferTo(file);
-            }catch (IOException e){
+            } catch (IOException e) {
                 e.printStackTrace();
             }
         }
+
         //其他数据
+        //原商品
         cn.bps.pojo.Product productPojo = product.toProduct(id);
+        //编辑新商品的同时下架原商品
         Product newProduct = productService.editOne(productPojo);
-        Integer productId = productService.getRecentProductId(newProduct);
-        List<ProductBindFilter> productBindFilters = productBindFilterService.cloneByProductId(id,productId);
+        Integer newProductId = productService.getRecentProductId(newProduct);
+        //拷贝筛选标签给新商品
+        List<ProductBindFilter> productBindFilters = productBindFilterService.cloneByProductId(id, newProductId);
         productBindFilterService.insertProductBindFilter(productBindFilters);
 
-        productImageService.updateProductImage(productPojo.getId(),fileName);
+        if(filePath.equals("")){
+            String oldImg = productImageService.getImageUrl(productPojo.getId());
+            productImageService.addProductImage(newProductId,oldImg);
+        }
+        //拷贝新商品的
+        productImageService.addProductImage(newProductId, filePath);
 
-        return "redirect:/admin/product";
+        return "redirect:/admin/product?start="+id;
     }
 
     //提交添加的产品
     @RequestMapping(value = "/add/post", method = RequestMethod.POST)
     public String addProduct(HttpServletRequest request,
-                             @ModelAttribute CompleteProduct completeProduct)
-    {
+                             @ModelAttribute CompleteProduct completeProduct,
+                             @ModelAttribute("sessionUsername") String username) {
+
+        if (username.equals("")) return "redirect:/admin/login";
 
         cn.bps.pojo.Product product = completeProduct.generatorProduct();
         completeProduct.init();
@@ -148,19 +166,19 @@ public class AdminProductController {
 
         String filePath = null;
 
-        if(!text.equals("")){
-            try{
-                filePath = "/img/file/"+fileName;
-                File file = new File(request.getServletContext().getRealPath("/WEB-INF/img/file/"),fileName);
-                if(!file.exists()){
+        if (!text.equals("")) {
+            try {
+                filePath = "/img/file/" + fileName;
+                File file = new File(request.getServletContext().getRealPath("/WEB-INF/img/file/"), fileName);
+                if (!file.exists()) {
                     file.mkdir();
                 }
                 multipartFile.transferTo(file);
 
-            }catch (IOException e){
+            } catch (IOException e) {
                 e.printStackTrace();
             }
-        }else{
+        } else {
             filePath = "http://temp.im/200x200";
         }
 
@@ -176,10 +194,10 @@ public class AdminProductController {
 
     //查看商品详细信息
     @RequestMapping(value = "/info/{id}")
-    public String showProductInfoView(@PathVariable(value = "id")int id,
-                              Model model)
-    {
+    public String showProductInfoView(@PathVariable(value = "id") int id,
+                                      Model model, @ModelAttribute("sessionUsername") String username) {
 
+        if (username.equals("")) return "redirect:/admin/login";
 
         cn.bps.pojo.Product product = productService.getProductById(id);
 
@@ -187,19 +205,27 @@ public class AdminProductController {
 
 
         String imageUrl = productImageService.getImageUrl(product.getId());
-        model.addAttribute("image",imageUrl);
+        model.addAttribute("image", imageUrl);
 
         List<Category> categories = categoryService.getCategories();
-        model.addAttribute("categories",categories);
+        model.addAttribute("categories", categories);
+
+
+        List<Integer> filterIdList = productBindFilterService.getConcreteFilterIdsByProduct(product);
+
+        List<ConcreteFilter> concreteFilterList = concreteFilterService.getConcreteFilterByIds(filterIdList);
+
+        model.addAttribute("filters",concreteFilterList);
 
         return "admin/product/productInfo";
     }
 
 
     @RequestMapping(value = "/edit/{id}")
-    public String editProduct(@PathVariable(value = "id")int id,
-                                  Model model)
-    {
+    public String editProduct(@PathVariable(value = "id") int id,
+                              Model model, @ModelAttribute("sessionUsername") String username) {
+
+        if (username.equals("")) return "redirect:/admin/login";
 
         cn.bps.pojo.Product product = productService.getProductById(id);
 
@@ -214,26 +240,34 @@ public class AdminProductController {
     }
 
     @RequestMapping(value = "/del/{id}")
-    public String delProduct(@PathVariable(value = "id")int id)
-    {
+    public String delProduct(@PathVariable(value = "id") int id,
+                             @ModelAttribute("sessionUsername") String username) {
 
+        if (username.equals("")) return "redirect:/admin/login";
 
         //首先删除相关筛选
         cn.bps.pojo.Product product = productService.getProductById(id);
         productBindFilterService.deleteDemos(id);
 
-        int productId = productService.deleteOneById(id);
-
-        if(productId==id){
-
-            //删除相关图片
-            if(productImageService.deleteOneByProductId(id)>0)
-
+        if (productImageService.deleteOneByProductId(id) > 0){
+            int productId = productService.deleteOneById(id);
+            if (productId == id) {
                 return "redirect:/admin/manage";
+            }
         }
-
         //失败
         return "redirect:/admin/product";
     }
+
+
+    @ModelAttribute("sessionUsername")
+    public String BeforeAdminController(HttpSession session) {
+        String username = (String) session.getAttribute("username");
+        if (username != null) {
+            return username;
+        }
+        return "";
+    }
+
 
 }
